Create API test for creating a user
Also update test file to use transactions in order to
delete the created user after the test is complete.

Note: A new DBAL type was created to handle datetimes in SQL Server. It
was required due to the StackOverflow database using `datetime` columns
instead of `datetime2` columns. See here for more info:
https://www.doctrine-project.org/projects/doctrine-orm/en/3.2/reference/limitations-and-known-issues.html#microsoft-sql-server-and-doctrine-datetime